@@ -14,11 +14,12 @@
     "@types/uuid": "9.0.6",
     "axios": "1.6.5",
     "daisyui": "3.9.4",
-    "next": "13.5.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.1.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-intersection-observer": "9.5.2",
-    "uuid": "9.0.1"
+    "uuid": "9.0.1",
+    "yarn": "^1.22.22"
   },
   "devDependencies": {
     "@testing-library/dom": "9.3.3",
@@ -33,10 +34,10 @@
     "autoprefixer": "10.4.16",
     "babel-jest": "29.7.0",
     "eslint": "^8.52.0",
-    "eslint-config-next": "13.5.6",
+    "eslint-config-next": "^14.1.4",
     "jest": "29.7.0",
     "postcss": "8.4.31",
-    "tailwindcss": "3.3.3",
+    "tailwindcss": "3.4.3",
     "ts-jest": "29.1.1",
     "typescript": "5.2.2"
   },
