@@ -1,5 +1,8 @@
 module.exports = {
-    purge: ["./pages/**/*.{js,ts,jsx,tsx}", "./components/**/*.{js,ts,jsx,tsx}"],
+    content: [
+        "./app/**/*.{js,ts,jsx,tsx}",
+        "./components/**/*.{js,ts,jsx,tsx}"
+    ],
     darkMode: 'media', // or 'media' or 'class'
     theme: {
         extend: {},
